Add size selection and refactor
